backend/pyproject.toml
[tool.poetry]
name = "lektprojekt-server"
version = "0.1.0"
description = "LektProjekt Application Server"
authors = ["Paweł Sacawa <psacawa@math.toronto.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
django-allauth = "^0.44.0"
django-configurations = "^2.2"
django-extensions = "^3.1.0"
django-filter = "^2.4.0"
django-split-settings = "^1.0.1"
djangorestframework = "^3.12.2"
colored-traceback = "^0.3.0"
dj-rest-auth = "^2.1.2"
psycopg2 = "^2.8.6"
drf-yasg = "^1.20.0"
django-model-utils = "^4.1.1"
django-polymorphic = "^3.0.0"
spacy = "^2.3.5"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
ipython = "^7.19.0"
progress = "^1.5"
tabulate = "^0.8.7"
pre-commit = "^2.9.3"
mypy = "^0.790"
pylint = "^2.6.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
# multi_line_output = 3
